fixed bigint cast

+++ api/lib/pg/directory/index.js
@@ -63,7 +63,7 @@ export async function listAllDirectories(usernameOrEmail, parent_directory_uuid)
           d."name",
           d."timestamp",
           d."download_count",
-          SUM(f."size")::integer AS "size",
+          SUM(f."size")::bigint AS "size",
           max(f."clamav_status") AS "clamav_status"
         FROM private."directory" d
           LEFT JOIN private."user" u ON (u."uuid" = d."user_uuid")
